backend/requirements.txt
Flask==1.1.2
python-dotenv==0.14.0
flask-wtf==0.14.3
opencv-python==4.4.0.40
Pillow==7.2.0
flask-cors==3.0.9
pyyaml==5.3.1
gunicorn==20.0.4